+++ src/c_console.c
@@ -753,6 +753,8 @@ void C_ShowConsole(void)
 
     if (gamestate == GS_TITLESCREEN && !devparm)
         S_StartSound(NULL, sfx_swtchn);
+    else if (gamestate == GS_LEVEL)
+        S_SetMusicVolume(musicVolume * MAX_MUSIC_VOLUME / 31 / 2);
 
     SDL_StartTextInput();
 }
@@ -770,6 +772,8 @@ void C_HideConsole(void)
         consoleactive = false;
         S_StartSound(NULL, sfx_swtchx);
     }
+    else if (gamestate == GS_LEVEL)
+        S_SetMusicVolume(musicVolume * MAX_MUSIC_VOLUME / 31);
 }
 
 void C_HideConsoleFast(void)