+++ releasenotes.md
@@ -1,6 +1,6 @@
 ### DOOM Retro v3.5.5
 
-* Music is now quieter when the menu is open and a game is being played.
+* Music is now quieter when the menu or console is open and a game is being played, or the game is paused.
 
 ---